Put string-version in hashmap for intent extras that are not serializable
When an app supports deep linking, a URI intent extra is sent along to the
app (key is android.intent.extra.REFERRER). The error I got when trying to
put this particular extra in the hashmap was:
"Exception in HostObject::getpropName:LaunchArguments):
java.lang.IllegalArgumentException: Could not convert class
android.net.Uri$StringUri"
Checking whether the extra is serializable first, and calling getString instead
of string if this is the case, solves this crash.